requirements.txt
mem0ai
openai
